Refactoring ast and interpreter to use visitor pattern for evaluating, confirmed tests pass
