+++ src/interpreter/mod.rs
@@ -1,4 +1,5 @@
 pub mod interpreter;
 pub mod errors;
+pub mod visitor;
 
 pub use interpreter::*;
